+++ module/ldbc-renewal-schema/src/main/scala/ldbc/schema/Table.scala
@@ -11,7 +11,7 @@ import scala.deriving.Mirror
 
 import ldbc.dsl.codec.Decoder
 import ldbc.statement.{ AbstractTable, Column }
-import ldbc.schema.interpreter.Tuples
+import ldbc.schema.interpreter.*
 
 trait Table[T](val $name: String) extends AbstractTable[T]:
 
@@ -25,20 +25,6 @@ trait Table[T](val $name: String) extends AbstractTable[T]:
   override def toString: String = s"Table($$name)"
 
 object Table:
-
-  /** A type function to pull a type parameter it has from a type with one type parameter. */
-  type Extract[T] = T match
-    case Option[t] => Extract[t]
-    case Array[t]  => Extract[t]
-    case List[t]   => Extract[t]
-    case Seq[t]    => Extract[t]
-    case Set[t]    => Extract[t]
-    case _         => T
-
-  /** A type function that derives its type from the type parameters that Option has. */
-  type ExtractOption[T] = T match
-    case Option[t] => Extract[t]
-    case _         => T
 
   private[ldbc] case class Opt[T](
     $name:   String,
